wizard/requirements.txt
fastapi
pydantic
pyyaml
uvicorn